holy shit

+++ apps/customer/models.py
@@ -27,9 +27,9 @@ class Customer(models.Model):
     address = models.CharField(max_length=100, default='Tehran')
     phone = models.CharField(max_length=12, unique=False, default='')
     domestic_bank_account_id = models.CharField(unique=False, max_length=16, null=True)
-    rial_wallet = models.IntegerField(default=0)
-    dollar_wallet = models.IntegerField(default=0)
-    euro_wallet = models.IntegerField(default=0)
+    rial_wallet = models.FloatField(default=0)
+    dollar_wallet = models.FloatField(default=0)
+    euro_wallet = models.FloatField(default=0)
 
     def __str__(self):
         return self.first_name
